third_party/blink/web_tests/external/wpt/css/css-anchor-position/position-fallback-tree-scoped-expected.txt

This is a testharness.js-based test.
PASS Document position-fallback matches @position-fallback in document scope
PASS Document position-fallback does not match @position-fallback in #outer_host scope
PASS Document position-fallback does not match @position-fallback in #inner_host scope
PASS Outer position-fallback matches @position-fallback in document scope
FAIL Outer position-fallback matches @position-fallback in #outer_host scope assert_equals: expected 200 but got 0
PASS Outer position-fallback does not match @position-fallback in #inner_host scope
PASS Inner position-fallback matches @position-fallback in document scope
FAIL Inner position-fallback matches @position-fallback in #outer_host scope assert_equals: expected 200 but got 0
FAIL Inner position-fallback matches @position-fallback in #inner_host scope assert_equals: expected 300 but got 0
Harness: the test ran to completion.
